\name{cvriskMSM}
\alias{cvriskMSM}
\title{Cross-validation for Boosting Multi-state Models}
\description{Cross-validation for Boosting Multi-state Models.}
\usage{
cvriskMSM(m, d, id, formulaMSM, xlist, qlist, k, riskset)
}
\arguments{
\item{m}{...}
\item{d}{...}
\item{id}{...}
\item{formulaMSM}{...}
\item{xlist}{...}
\item{qlist}{...}
\item{k}{...}
\item{riskset}{...}
}
\details{...}
\value{...}
\author{Holger Reulen}