\name{meancentering}
\alias{meancentering}
\title{Mean Centering with Taking Care of the Transition Type(s)}
\description{...}
\usage{meancentering(d, x, q, x.name, q.name)}
\arguments{
  \item{d}{data set}
  \item{x}{covariate}
  \item{q}{transition type(s)}
  \item{x.name}{name of the covariate for pasting the new transition
    type specific covariate name}
  \item{q.name}{name of the transition type for pasting the new transition
    type specific covariate name}
}
\details{...}
\value{...}
\author{Holger Reulen}