\name{plotcvriskMSM}
\alias{plotcvriskMSM}
\title{Plot Cross-validation for Boosting Multi-state Models}
\description{Plot cross-validation for boosting multi-state models.}
\usage{
plotcvriskMSM(cvriskMSMobject, type)
}
\arguments{
\item{cvriskMSMobject}{result from \code{cvriskMSM}}
\item{type}{should all stratified results be plotted (\code{"all"}, default), 
or only mean (\code{"mean"})}
}
\details{...}
\value{...}
\author{Holger Reulen}