\name{toolbox}
\alias{toolbox}
\title{Toolbox for gamboostMSM package}
\description{This function implements prctical tools for 
  estimation of multi state models using gamboostMSM package.}
\usage{
toolbox(tool, input)
}
\arguments{
\item{tool}{ a character string defining the tool to use}  
\item{input}{ a list of input objects for the specific tools.}
}
\details{See users manual for how to use the toolbox.}
\value{ output object of the used tools.}
\author{Holger Reulen}